wording
which is weird because I don't reread these entries.

I would think it's a great way to look at my brilliance, but the "image"
in my head is far brighter and much fuller than what I wrote.
I prefer the image in my head to what I wrote because what I wrote is a
2d shadow of a 3d idea