add on to my perspective.
after sending those emails, I've started rereading things. it's
interesting to see my own thoughts and the reactions to them.

I'll probably start rereading and updating some ideas a bit more to try
and take, on marc's playstyle for a bit and see how it gees.

the replay is a very important learning tool.